--- src/System.Private.CoreLib/src/Internal/Reflection/Tracing/ReflectionTrace.Internal.cs
@@ -1,17 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-using System.Diagnostics;
-
-using ReflectionTraceCallbacks = Internal.Runtime.Augments.ReflectionTraceCallbacks;
-
-namespace Internal.Reflection.Tracing
-{
-    internal static partial class ReflectionTrace
-    {
-        private static ReflectionTraceCallbacks s_callbacks;
-    }
-}
-
