+++ src/System.Private.Reflection.Core/src/Internal/Reflection/Tracing/ReflectionTrace.Internal.cs
@@ -148,7 +148,7 @@ namespace Internal.Reflection.Tracing
             if (!type.IsRuntimeImplemented())
                 return null;
 
-            RuntimeTypeInfo runtimeType = type.GetRuntimeTypeInfo<RuntimeTypeInfo>();
+            RuntimeTypeInfo runtimeType = type.CastToRuntimeTypeInfo();
             if (runtimeType.HasElementType)
             {
                 String elementTypeName = runtimeType.InternalRuntimeElementType.CastToType().NonQualifiedTypeName();
@@ -233,7 +233,7 @@ namespace Internal.Reflection.Tracing
             if (!type.IsRuntimeImplemented())
                 return null;
 
-            RuntimeTypeInfo runtimeType = type.GetRuntimeTypeInfo<RuntimeTypeInfo>();
+            RuntimeTypeInfo runtimeType = type.CastToRuntimeTypeInfo();
             if (runtimeType == null)
                 return null;
             String nonqualifiedTypeName = runtimeType.CastToType().NonQualifiedTypeName();
@@ -250,7 +250,7 @@ namespace Internal.Reflection.Tracing
             if (!type.IsRuntimeImplemented())
                 return null;
 
-            RuntimeTypeInfo runtimeTypeInfo = type.GetRuntimeTypeInfo<RuntimeTypeInfo>();
+            RuntimeTypeInfo runtimeTypeInfo = type.CastToRuntimeTypeInfo();
             if (runtimeTypeInfo is RuntimeNoMetadataNamedTypeInfo)
                 return null;
             return runtimeTypeInfo.Assembly.NameString();
