+++ src/System.Private.Reflection.Core/src/System/Reflection/Runtime/TypeInfos/RuntimeGenericParameterTypeInfoForMethods.cs
@@ -78,7 +78,7 @@ namespace System.Reflection.Runtime.TypeInfos
         {
             get
             {
-                TypeContext typeContext = this.DeclaringType.GetRuntimeTypeInfo<RuntimeTypeInfo>().TypeContext;
+                TypeContext typeContext = this.DeclaringType.CastToRuntimeTypeInfo().TypeContext;
                 return new TypeContext(typeContext.GenericTypeArguments, DeclaringRuntimeNamedMethodInfo.RuntimeGenericArgumentsOrParameters);
             }
         }
