+++ src/System.Private.Reflection.Core/src/System/Reflection/Runtime/TypeInfos/RuntimeGenericParameterTypeInfoForTypes.cs
@@ -18,7 +18,7 @@ namespace System.Reflection.Runtime.TypeInfos
 {
     internal sealed partial class RuntimeGenericParameterTypeInfoForTypes : RuntimeGenericParameterTypeInfo
     {
-        private RuntimeGenericParameterTypeInfoForTypes(MetadataReader reader, GenericParameterHandle genericParameterHandle, RuntimeNamedTypeInfo declaringRuntimeNamedTypeInfo)
+        private RuntimeGenericParameterTypeInfoForTypes(MetadataReader reader, GenericParameterHandle genericParameterHandle, RuntimeTypeInfo declaringRuntimeNamedTypeInfo)
            : base(reader, genericParameterHandle)
         {
             DeclaringRuntimeNamedTypeInfo = declaringRuntimeNamedTypeInfo;
@@ -52,7 +52,7 @@ namespace System.Reflection.Runtime.TypeInfos
             }
         }
 
-        internal RuntimeNamedTypeInfo DeclaringRuntimeNamedTypeInfo { get; }
+        internal RuntimeTypeInfo DeclaringRuntimeNamedTypeInfo { get; }
     }
 }
