+++ src/System.Private.Reflection.Execution/src/System.Private.Reflection.Execution.csproj
@@ -16,9 +16,6 @@
   <PropertyGroup Condition="'$(IsProjectNLibrary)' != 'true'">
     <DefineConstants>CORERT;$(DefineConstants)</DefineConstants>
   </PropertyGroup>
-  <PropertyGroup Condition="'$(IsProjectNLibrary)' == 'true'">
-    <DefineConstants>ENABLE_REFLECTION_TRACE;$(DefineConstants)</DefineConstants>
-  </PropertyGroup>
 
   <!-- Setup the right references -->
   <ItemGroup>
@@ -76,7 +73,6 @@
     <Compile Include="Internal\Reflection\Execution\ExecutionEnvironmentImplementation.MetadataTable.cs" />
     <Compile Include="Internal\Reflection\Execution\ExecutionEnvironmentImplementation.ManifestResources.cs" />
     <Compile Include="Internal\Reflection\Execution\ReflectionExecutionDomainCallbacksImplementation.cs" />
-    <Compile Include="Internal\Reflection\Execution\ReflectionTraceCallbacksImplementation.cs" Condition="'$(IsProjectNLibrary)' == 'true'" />
     <Compile Include="Internal\Reflection\Execution\MetadataReaderExtensions.cs" />
     <Compile Include="Internal\Reflection\Execution\MetadataNameExtensions.cs" />
     <Compile Include="Internal\Reflection\Execution\EnumInfoImplementation.cs" />
